pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stem-cell-code-auditor"
version = "0.1.0"
dependencies = [
    "gitpython",
    "langchain",
    "langchain-openai",
    "langgraph",
    "dotenv",
]

[project.scripts]
run-me = "stem_cell_code_auditor.main:main"
